Simplify and rename Python SDK LLM callbacks to CopilotRequestHandler
Mirror the .NET/Node simplification + terminology rename in the Python SDK:
consolidate the provider/handler two-layer design into a single
copilot/copilot_request_handler.py, drop the accepted:false ack plumbing and
the staged backstop, emit the WebSocket 101 upgrade head eagerly (a lazy
bridge deadlocks the runtime connect), and rename the public Llm* types to
Copilot* (types carry the prefix; attributes/methods stay succinct). The
client option becomes request_handler: CopilotRequestHandler. Generated wire
types are untouched.

Stream in-memory httpx responses (built with content=) by forwarding their
buffered body, since their raw stream is already consumed and cannot be
re-iterated; real streamed responses still flow through aiter_raw.

Consolidate the e2e suite to three files (test_copilot_request_handler
covering HTTP + WebSocket + streaming, test_copilot_request_session_id, and
test_copilot_request_cancel_error with the error and runtime-cancel cases)
plus a shared helpers module, replacing the eight test_llm_inference_* files.
